Fixed nullability stuff and make IsRootElementLoaded public
